Filtering Element type for LCOM and LOC aggregators
As the metamodel is supplemented with more Element types, the aggregators should be changed to handle these additional types. For both aggregators a filter is added that only lets the types, interesting for the specific aggregator, pass. This will avoid an incorrect measurement of the metric further up in the chain.